firestrike_temp/requirements.txt
stem==1.8.1
PySocks==1.7.1
cryptography==41.0.3
pynacl==1.5.0
aiohttp==3.8.5
async-timeout==4.0.3
attrs==23.1.0 